requirements.txt
asttokens==2.4.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
dnspython==2.7.0
executing==2.1.0
experta==1.9.4
Flask==3.1.1
flask-cors==6.0.0
fonttools==4.54.1
frozendict==1.2
gunicorn==23.0.0
idna==3.10
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==11.0.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
pyasn1==0.6.0
pygame==2.6.0
Pygments==2.18.0
pymongo==4.13.0
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.1
pywin32==306
pyzmq==26.2.0
requests==2.32.3
rsa==4.9
schema==0.6.7
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.4.1
traitlets==5.14.3
tzdata==2024.1
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
